TGDB Site/API Search Enhancements
Completed: Issue #18 - When performing a title search using the API (GetGamesList.php), Alt-Titles are also included in the search

Completed: Issue #22 - Changed the mysql fulltext keyword length to "1" so that titles such as "X-men" and "Quake 4" are matched and ranked correctly

-----------------
Additional Notes:
-----------------

These improvements affect both the "In-site HARD search results" and the API search results.

There are some significant improvements to the site search which I have been able to achieve, the most notable of which are:

Improved Relevance Ordering - Titles returned are now more relevant to the initial search query (with non generic identical matches appearing top of the list in *MOST* cases)
Numeric Values/Tokens Are Now Taken Into Account When Searching - For example a search for "Quake II" actually looks for the "II" and orders the numerous exact matches to the original search string first
Most Special Characters Will Not Break the Search - Searching for games like "Call Of Duty: Modern Warfare 2" will find the correct results when using the search string "call of duty modern warfare 2" <- i.e without the colon
Alternate Titles are Now Searched Upon - When Searching for games that may have an alt title such as "Call Of Duty: Modern Warfare 4" which is sometimes referred to and may have the Alt. Title of "MW2", the game will still be found

But, of course, like the current search system, there are some remaining issues and imperfections that I fear may not be able to be resolved with our current system/set-up.

These sometimes include (when performing certain searches):

When a match is found for an alt title, in some cases, due to restrictions of what is possible with our simple mysql server, the match may appear further down the list of results than we would like.
With some searches that incorporate large search strings, there are sometimes extra results that are loosely relevant to the search term, but thankfully these have a tendency to appear at the end of the results list (are lower ranked matches).

==========
IMPORTANT:
==========
The MySQL Fulltext Min-Char Length parameter on the Live server requires to be set to "1" for this updateto function as optimally as possible.