Fixes from @chaeril 's pull request
Fixed: @chaeril had removed the facebox modal plugin includes from the
head of the 'templates/front.php' file causing the 'share via email'
functionality to break.
Fixed: Moved the Facebook Javascript API library include to bottom of
the "templates/footer.php" file as it would be best loaded after the
rest of the page content so as not to impact performance, and also
because it contained html that should not be inserted in the <head>
section of a document, thus preventing the Facebook share button from
showing.