Zabanya Recent ClearLogo's Fixes
Fixed: ClearLOGO's were being loaded at full size: changed image width
to 300px.
Fixed: Broken conditional statement was causing both the original
clearlogo image and a non-existant cahced copy to be rendered into the
html, causing missing image icon boxes to be displayed on some browsers.
Add an "else" to the conditional so that only one or the other is
displayed.
Modified: Changed the order of Art Report links on the Stats Page.