Duplicate Games Prevention Methods
Completed: Issue #124 and Issue #125 - Taken methods to decrease the
amounts of duplicate games being submitted on the site. Measures taken
include modification of the Add Game page to display a list of currently
existing games relevant to the desired game being added, and asking the
user to confirm whether they can see an existing game that matches the
one they were trying to add. Also, if users forcibly add a game to a
platform, and an exact match is found in the DB before adding, they are
instead re-directed the the existing game page.
Additional: Updated styling of messages and error messages.
Additional: Modified the AJAX elasticsearch script to also accept a
platform parameter to filter by.