Provide a VM for tests, using vagrant and ansible.
This commit adds a Vagrantfile that brings up the standard ubuntu/trusty64
vagrant virtualbox VM, provisions it with the required apache, PHP and
MySQL packages and configurations, and installs the site. To do that,
install vagrant on the host and do a "vagrant up" in the working copy
of the repository. The site will be available at http://localhost:8888/