Provide the 'exactname' api/GetGame.php parameter.
The 'exactname' parameter allows the user to specify the exact name
of the game. The implementation uses the 'exactname' parameter to build
a SQL with '... WHERE GameTitle = $exactname' instead of using
MATCH and LIKE statements.

The overall effect is that games with titles such as '20 em 1' and
'Alf' that can't be retrieved with the 'name' parameter can now be
retrieved by using 'exactname'